Change guid to uint32 thanks to sharm for pointing out

+++ src/game/Guild.h
@@ -244,7 +244,7 @@ struct MemberSlot
         return false;
     }
     uint32 accountId;
-    uint64 guid;
+    uint32 guid;
     uint64 logout_time;
     std::string name;
     std::string Pnote;
@@ -287,7 +287,7 @@ class Guild
         void Disband();
 
         typedef std::map<uint32, MemberSlot> MemberList;
-        typedef std::list<uint64> MemberGuidList;
+        typedef std::list<uint32> MemberGuidList;
         typedef std::vector<RankInfo> RankList;
 
         uint32 GetId(){ return Id; }
